# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Coral Companion API Documentation",
        "description": "Documentation for the API used in the Coral Companion app",
        "contact": {
            "name": "Paula Rumeu",
            "email": "paururo08@hotmail.com"
        },
        "license": {
            "name": "Creative Commons Attribution Share Alike 4.0 International",
            "url": "https://creativecommons.org/licenses/by-sa/4.0/"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://github.com/salem404/coral-companion/raw/master/frontend/src/assets/img/logo-color.svg"
        }
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Localhost"
        },
        {
            "url": "https://coral-companion.railway.com/api",
            "description": "Railway"
        }
    ],
    "paths": {
...
        "/profiles": {
            "get": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Get all profiles",
                "description": "Get all profiles from the database",
                "operationId": "1afd045c87391eb70de9615a4e03cf91",
                "responses": {
                    "200": {
                        "description": "Success: Return all profiles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Profile"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No profiles found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No profiles found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Create a profile",
                "description": "Create a new profile using the data provided in the request body. The user must be an admin to create a profile for other users. An user can create a profile for themselves (user_id not required).",
                "operationId": "14ca107420d028924052574e1d385c1c",
                "requestBody": {
                    "$ref": "#/components/requestBodies/ProfileCreate"
                },
                "responses": {
                    "201": {
                        "description": "Success: A profile has been created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profile created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request: The request you sent was invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "farmer_name, farm_name and color are required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Only admins or the user themselves can create a profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You are not authorized to create a profile"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/profiles/{id}": {
            "get": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Get a profile",
                "description": "Get a profile by ID in the URL",
                "operationId": "9aafb9b7a4b31cb41fa18c41e18dd6e0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of profile",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: Returns a profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Profile"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found: Profile doesn't exist",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profile with id 4 not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Update a profile",
                "description": "Update a profile by ID in the URL and data provided in the request body. Admins can update any profile, users can only update their own profiles",
                "operationId": "fad658ba45fe7dab378d24861bf22f4c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of profile",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/ProfileUpdate"
                },
                "responses": {
                    "200": {
                        "description": "Success: Profile has been updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profile updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request: Data validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The farmer name field is required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Only admins or the user who created the profile can update it",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You are not authorized to update this profile"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found: Profile doesn't exist",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profile with id 4 not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Delete a profile",
                "description": "Delete a profile by ID in the URL. Admins can delete any profile, users can only delete their own profiles",
                "operationId": "b3f355620753ead8a083666cdd71362c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of profile to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: Profile deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profile deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Only admins or the user who created the profile can delete it",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You are not authorized to delete this profile"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found: Profile doesn't exist",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profile with id 4 not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Profile": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 4
                    },
                    "farmer_name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "farm_name": {
                        "type": "string",
                        "example": "John's Farm"
                    },
                    "color": {
                        "type": "string",
                        "example": "#000000"
                    },
                    "user_id": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "required": [
                    "id",
                    "username",
                    "email",
                    "password"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 1
                    },
                    "username": {
                        "type": "string",
                        "example": "john_doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "example@email.es"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "password123"
                    }
                },
                "type": "object"
            }
        },
        "requestBodies": {
...
            "ProfileCreate": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "farmer_name",
                                "farm_name",
                                "color"
                            ],
                            "properties": {
                                "farmer_name": {
                                    "type": "string",
                                    "example": "John Doe"
                                },
                                "farm_name": {
                                    "type": "string",
                                    "example": "John's Farm"
                                },
                                "color": {
                                    "type": "string",
                                    "example": "#000000"
                                },
                                "user_id": {
                                    "type": "integer",
                                    "example": 1
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "ProfileUpdate": {
                "description": "Profile object that needs to be added to the database",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "farmer_name": {
                                    "type": "string",
                                    "example": "John Doe"
                                },
                                "farm_name": {
                                    "type": "string",
                                    "example": "John's Farm"
                                },
                                "color": {
                                    "type": "string",
                                    "example": "#000000"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}